Effective Sampling Rate

Calculates the effective sampling rate (esr).

# S3 method for mcmcr
esr(x, by = "all", as_df = FALSE, na_rm = FALSE, ...)

Arguments

x

An object.

by

A string indicating whether to determine by "term", "parameter" or "all".

as_df

A flag indicating whether to return the results as a data frame versus a named list.

na_rm

A flag specifying whether to ignore missing values.

...

Other arguments passed to methods.

Value

A number between 0 and 1 indicating the esr value.

Details

By default $$\frac{1}{1 + 2 \sum_{k = 1}^\infty\rho_k(\theta)}$$ from Brooks et al. (2011) where the infinite sum is truncated at lag \(k\) when \(\rho_{k+1}(\theta) < 0\).

References

Brooks, S., Gelman, A., Jones, G.L., and Meng, X.-L. (Editors). 2011. Handbook for Markov Chain Monte Carlo. Taylor & Francis, Boca Raton.

See also

Examples

esr(mcmcr_example)
#> [1] 0.011