Term coefficients

Gets coefficients for all the terms in an MCMC object.

# S3 method for mcmc
coef(object, conf_level = 0.95, estimate = median, simplify = FALSE, ...)

Arguments

object
The MCMC object to get the coefficients for

conf_level
A number specifying the confidence level. By default 0.95.

estimate
The function to use to calculate the estimate.

simplify
A flag specifying whether to return just the estimate, lower, upper and svalue.

...
Unused.

Value

An data frame of the coefficients with the columns indicating the term, estimate, the standard deviation (sd) and zscore (if simplify = FALSE), lower and upper credible intervals and pvalue or svalue (if simplify = TRUE).

Methods (by class)

mcmc: Get coefficients for terms in mcmc object

See also

Examples

coef(mcmcr_example)
#> Warning: The `simplify` argument of `coef()` must be TRUE as of mcmcr 0.4.1.
#> This warning is displayed once every 8 hours.
#> Call `lifecycle::last_warnings()` to see where this warning was generated.
#> # A tibble: 7 × 7
#>   term      estimate    sd zscore lower upper  pvalue
#>   <term>       <dbl> <dbl>  <dbl> <dbl> <dbl>   <dbl>
#> 1 alpha[1]     3.72  0.901   4.15 2.21   5.23 0.00125
#> 2 alpha[2]     4.72  0.901   5.26 3.21   6.23 0.00125
#> 3 beta[1,1]    0.972 0.375   2.57 0.251  1.71 0.0237 
#> 4 beta[2,1]    1.97  0.375   5.24 1.25   2.71 0.00624
#> 5 beta[1,2]    1.97  0.375   5.24 1.25   2.71 0.00624
#> 6 beta[2,2]    2.97  0.375   7.91 2.25   3.71 0.00125
#> 7 sigma        0.791 0.741   1.31 0.425  2.56 0.00125